Split many methods from form and popup classes into children
