+++ src/js/components/Popup.js
@@ -10,10 +10,6 @@ export default class Popup extends BaseComponent {
     this.open = this.open.bind(this);
     this.close = this.close.bind(this);
     this.clearContent = this.clearContent.bind(this);
-    this.setContent = this.setContent.bind(this);
-    this._linkToSignin = this._linkToSignin.bind(this);
-    this._linkToSignup = this._linkToSignup.bind(this);
-    this._linkToRegistered = this._linkToRegistered.bind(this);
     this._checkUserEvents = this._checkUserEvents.bind(this);
   }
 
@@ -26,21 +22,6 @@ export default class Popup extends BaseComponent {
   clearContent() {
     this._unmount();
     this._clearNodeContent(this._domElement.querySelector(this._container));
-  }
-
-  _linkToSignin() {
-    this.clearContent();
-    this._dependecies.popupSignin.setContent();
-  }
-
-  _linkToSignup() {
-    this.clearContent();
-    this._dependecies.popupSignup.setContent();
-  }
-
-  _linkToRegistered() {
-    this.clearContent();
-    this._dependecies.popupRegistered.setContent();
   }
 
   open() {
@@ -64,28 +45,12 @@ export default class Popup extends BaseComponent {
   }
 
   /**
-   * Создаёт попап и устанавливает в него необходимое содержимое и обработчики
+   * Инициализация попапа, создаёт и добавляет стандартные обработчики
    */
-  setContent() {
+  _initPopup() {
     this._domElement.querySelector(this._container).appendChild(this._createPopup());
     this._mountLocalHandlers([{ element: this._closeElement, handlers: [this.close] },
       { element: document, handlers: [this._checkUserEvents], event: 'keydown' },
       { element: this._domElement, handlers: [this._checkUserEvents] }]);
-
-    if (this._props.popupName === 'popupSignup' && this._dependecies.popupSignin) {
-      this._mount({ element: '.popup__link', handlers: [this._linkToSignin] });
-
-      if (this._dependecies.signupForm) {
-        this._dependecies.signupForm.handlers();
-      }
-    } else if (this._props.popupName === 'popupSignin' && this._dependecies.popupSignup) {
-      this._mount({ element: '.popup__link', handlers: [this._linkToSignup] });
-
-      if (this._dependecies.signinForm) {
-        this._dependecies.signinForm.handlers();
-      }
-    } else if (this._props.popupName === 'popupRegistered' && this._dependecies.popupSignin && this._dependecies.signinForm) {
-      this._mount({ element: '.popup__link', handlers: [this._linkToSignin] });
-    }
   }
 }
